The store filter counts to sixty-two
Slice 1b part three, and the last ceiling. StoreBuilding.AllowedGoods was an
int bitmask with one bit per good and the Spoken sentinel at bit 30, so good
30 would have set the sentinel and a store the player never touched would
report that they had. Not a crash - a filter that switches itself on.

It is a 64-bit mask now with the sentinel at 62, and both halves are hashed:
mixing only the low word would have left every good above the thirty-second,
and the sentinel itself, outside the hash. That is the quiet kind of
determinism bug - two runs that read identical and are not.

The goldens did not move, and that is the guard's doing rather than luck. The
mask is hashed only when non-zero, and zero means "the player has not said",
which is every store in an unattended fifty-year run. The comment on that
guard predicted exactly this when filters landed.

The compiler caught a third bug on the way: mask &= ~bit, with bit still an
int, sign-extends when widened and wipes the sentinel. CS0675. That one would
have read as a filter forgetting it had been set.

Recorded, not fixed: a villager cannot carry stone. CarriedLogs, CarriedFirewood
and CarriedFood are three named fields, so of six goods three can be carried and
three cannot, and the clearing path spends the tile before discovering it has
nowhere to put the yield. It looks reachable rather than latent, but looks is
not measured - it needs a failing test first, and the fix moves the goldens.
goods-catalog.md 4.0, and a task chip is open for it.

Slice 2 is still outstanding and the work is unproven without it: a test that
defines a seventh good entirely in data. No new good ships into the game - the
proof is the test, not content nobody asked for.

740 passing, 0 failing, 2 skipped of 742, goldens unmoved.

+++ src/Bclone.Sim/Core/SimWorld.cs
@@ -3028,14 +3028,14 @@ public sealed class SimWorld
         // Start from "everything this kind can hold", so the first click narrows rather than
         // wiping. A mask of zero means no opinion, and turning one good off has to leave the
         // others on — otherwise the first click would empty the building.
-        int mask = store.AllowedGoods;
+        long mask = store.AllowedGoods;
         if (mask == 0)
         {
             for (int g = 0; g < GoodsCatalog.Count; g++)
             {
                 if (store.CanEverHold((Goods)g))
                 {
-                    mask |= 1 << g;
+                    mask |= 1L << g;
                 }
             }
         }
@@ -3045,7 +3045,7 @@ public sealed class SimWorld
         // opinion", and the store silently starts accepting everything again.
         mask |= StoreBuilding.Spoken;
 
-        int bit = 1 << (int)goods;
+        long bit = 1L << (int)goods;
 
         if (accepted)
         {
